man: remove old ordering discussion in fi_tagged.3
The notes about ordering of tagged operations in the
man page was older and not updated after we moved
ordering flags to fi_endpoint.3.

If we don't remove this text, it implies that users of
tagged operations cannot relax the ordering.

FI_ORDER_SAS as defined in fi_endpoint.3 should satisfy MPI ordering
requirements.

Signed-off-by: Sayantan Sur <sayantan.sur@intel.com>