mr: more general memory registration on fid
Currently memory registration can only be called on a domain object.

This patch removes this restriction and makes memory registration
available for any FI object. For example, memory may be registered on a
fabric object, endpoint object and so on.  This will keep OFI more
generic and flexible.

The static inline for memory registration by the apps remains unchanged.

Signed-off-by: Sayantan Sur <sayantan.sur@intel.com>