Add writedata support to RMA examples
Add a command-line option to support using the writedata operation to
send remote CQ data to the simple/fi_msg_rma and simple/fi_rdm_rma
tools.

We test that provider gives a sensible context value on the remote CQ
data completion.  The two sensible values are NULL (meaning that the
remote CQ data did not consume a recv WQE) or the context value of the
next queued fi_recv* (meaning that the remote CQ data consumed a recv
WQE).  We also test that the completion data is correctly transferred.

We add an enum to specify the RMA operation type, which allows us to
cleanly add writedata to the set of operations supported by the RMA
tests (since libfabric itself doesn't supply a constant that can be
cleanly used for this purpose).

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>