fabric: Allow use of fi_dupinfo to allocate an empty fi_info
Allow a user to call fi_dupinfo with a NULL source fi_info.
This will result in fi_dupinfo allocating an empty fi_info structure
and related substructures that the app may then fill in.  This
simplifies things for the app, so that it doesn't need to declare
a bunch of independent structures and link them together.

Provide a static inline wrapper around this called fi_allocinfo,
to make things easier on the user.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>