fabric: Add err_data_size field to fi_eq_err_entry
An application that calls fi_reject can pass user data as part
of the operation.  This data is available on the remote side as
part of an EQ error event (ECONREFUSED).  However, the recipient
does not know how much data referenced by the err_data pointer is
valid.  Add a data size field to the fi_eq_err_entry that an app
can use when reading any err_data that is intended to be accessed
directly by the application.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>