prov/psm: name server tweaks
A few changes related to the mini name server:

(1) make the name server thread detached and let it exit together
with the main thread. The previous way of using pthread_cancel and
pthread_join had caused memory management issue with SHMEM. This
also simplifies the code.

(2) move the creation of the name server thread from domain
opening to fabric opening. Also move the MPI environement setting
at the same time.

(3) a few adjustment to the domain and fabric data structure to
facilitate the above changes.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>