fabric: Add av_type field to domain attribute
Providers are currently required to support both AV tables and
maps on the same domain.  Although that is still possible, forcing
this could lead to implementation issues.  For example, a CQ
must be prepared to handle endpoints bound to either type.
Optimizations may be possible if the provider knows what type of
addressing the application requires up front.

Add an av_type field as a domain attribute.  Applications can
provide this as input regarding the type of AV that they intend
to use.  And providers can indicate if a specific AV type is
required.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>