man pages: move generated nroff pages to man/manX/
This commit keeps the Markdown man pages in `man/`, but moves the
generated nroff pages to `man/manX/`, where X is the section number of
the man page.  The helper scripts are updated to match this scheme.

Minor note: the md2nroff.pl change looks a little bigger than it
really is.  To output to a `man/manX/` subdir, I needed to morph the
$target_arg value, but I didn't want to change the semantic menaning
of $target_arg (i.e., it is the value that came in from the command
line).  So I moved the result to $target, but therefore had to rename
$target_arg -> $target all throughout the rest of the script.

Fixes #702.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>