prov/sockets: Fix 32-build warnings and crash
The sockets provider generates multiple build warnings and crashes
when running on a 32-bit system.  Fix the warnings and adjust the
code to handle 32-bit pointers, rather than assuming that the size
of all pointers is 64-bits.

To fix the 32-bit issues required performing a fairly low-level
review of the code.  Optimistic updates were made to the code to
align with coding guidelines, improve readability, etc.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>