prov/sockets: Fix hang when remote peer(s) close
Disable outbound buffering of data in order to pass the
fi_msg_pingpong and fi_rdm_pingpong tests.

When outbound buffering is enabled, the pingpong tests may fail.
The cause is that the server sends a reply, which gets buffered,
reads the completion for the reply, then exits.  The buffered
data can end up being lost, as it never gets flushed to the underlying
socket.

Flushing the data to the socket allows the kernel linger functionality
to push out all data until it has been acked.

The offending code is merely commented out for now.  Once the API
defines the correct behavior that the apps should use, it may be
possible to re-enable the buffering.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>