fabric: Redefine FI_REMOTE_COMPLETE
The current completion model is best described as a local
completion.  A CQ entry indicates that the local buffer may be
reused, and that is all.  It turns out that this is insufficient
for an application to ever know if its last data transfer was
ever received by the remote side.  This problem shows up in when
running pingpong tests over the sockets provider.  After the app
completes its last transfer, it exits.  As a result, data queued
by the provider is discarded and never received at the remote
side.

In order to support fast local completions, such that buffers can
re-used quickly, but also 'finalize' shutdown semantics that give
an app some hope that its data has been seen by a remote peer,
change the definition of FI_REMOTE_COMPLETE to indicate that a
completion should only be generated if the remote side will
receive the posted data (in the absence of errors).  This
definitions aligns with both IB and iWarp definitions for
completions.

The current meaning of FI_REMOTE_COMPLETE will be moved to a new
flag, FI_REMOTE_COMMIT.  That flag will indicate that the data has
been processed at the remote side and placed into the final
memory location.  This feature was requested by several apps, but
is not supported by current HW, only HW/SW combinations.

Because support for FI_REMOTE_COMPLETE is essential for proper
application behavior, it is removed as an optional capability,
and is instead required.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>