core: add FI_PROVIDER selection env var
FI_PROVIDER allow a user to restrict the set of providers that will be
returned by fi_getinfo().  The variable is a comma separated list of
provider names.  The selection is "inclusive" by default, but can be
made "exclusive" by prefixing the list with a "^" character.  The
provider names are not checked for validity, so specifying
`FI_PROVIDER=bogus,usnic` will select the "usnic" provider without any
warnings about "bogus".

Signed-off-by: Dave Goodell <dgoodell@cisco.com>