usnic: add --with-libnl configure option
@bturrubiates discovered recently that the usnic provider would not
build if you had a manually-installed version of libnl (v1 or v3) in a
non-standard location (i.e., outside of /usr).

After much discussion, it seems that the only solution is to add a
CLI option to configure:

* `--with-libnl(=dir)`: search for libnl v1 and v3 support, and abort
  configure if it can't be found.  If dir is specified, look for libnl
  support in that tree (i.e., include, and lib or lib64)

The negative of this switch can be specified, too, to disable support
for libnl v1 and/or libnl v3 (i.e., --without-libnl).  This
effectively disables the usnic provider (since the usnic provider
needs libnl support).

I also took this opportunity to document all the sadness / colloquial
knowledge that we have gained about libnl (v1 and v3) in comments in
prov/usnic/configure.m4.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>