prov/psm: turn on tagged messaging based RMA by default
The bandwidth of the RMA ops over the active message mechanism is
limited by the maximum payload size of an active message request.
An optimization is to use tagged messaging for large size RMA
operations. The trade off is that the number of available tag
bits is reduced, and that's the main reason the optimization was
off by default in the past.

No change the default setting to on in favor of better out-of-box
performance instead of available tag bits. The option can still
be turned off by setting OFI_PSM_TAGGED_RMA=0.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>