fabric: $FI_PROVIDER_PATH env var specifies where DLs live
Having a default location for DLs is fine (i.e., $pkglibdir), but it's
not enough.  Specifically, if a provider maintainer wants to provide
an asynchronous update to their built-in/inbox provider, it's easy
enough to create a DL of the provider and install it in $pkglibdir.

But what about the provider's extension header file?

That extension header file is currently installed into
$includedir/fi_ext_<provider>.h.  It would be extremely anti-social
for an async provider update to overwrite that file (and potentially
make package managers -- such as rpm -- unhappy).  Indeed, the point
of allowing DLs to "override" built-in providers is to provide async
provider updates *without* replacing any of the libfabric packaged
files (including the inbox-provided fi_ext_<provider>.h).

The async provider update will need to install its updated
fi_ext_<provider>.h file elsewhere -- e.g.,
/opt/vendor/include/fi_ext_<provider>.h.  ...but then it's weird that
the provider DL is in one location and the provider extension header
is in another (e.g., /usr/lib vs. /opt/vendor).  It would be nicer if
both could be installed under a single tree (e.g., /opt/vendor).

This commit therefore adds the FI_PROIVDER_PATH env var which allows a
colon-delimited list of directories to search for provider DLs.  If
$FI_PROVIDER_PATH is set, it is used instead of $pkglibdir.  If
$FI_PROVIDER_PATH is not set, $pkglibdir is used (just like it used to
be).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>