Makefile.am: log.c must be in $(common_srcs) or DL builds fail
It's an unfortunate model that was chosen (that we compile the same
common sources multiple times -- once in each provider), but move
log.c into $(common_srcs) to conform to that model.  This fixes the
problem of DL builds running into problems with missing symbols (the
fi_log_* symbols, in particular).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>