prov/verbs: Detect inject_size in fi_getinfo()
This allows the user to supply an inject_size that they would like to
use via tx_attr in hints to fi_getinfo(), and then have that value used
when creating the queue pair.

As written, this currently only works for active endpoints, since we
cannot create a queue pair corresponding to a passive endpoint.  For a
passive endpoint, we simply use our default guess or the value set by
the system administrator.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>