fabric: Define return value for EQ/CQ read of an empty queue
The return value is unclear from the man pages.  Clarify that
reading from an empty queue will return -FI_EAGAIN.  This
aligns (mostly) with the POSIX read/recv functionality when
dealing with nonblocking sockets. 

Also require that -FI_EAGAIN be returned even for the blocking
sread calls.  This keeps the behavior consistent between read and
sread, and avoids any confusion if sread is called with a timeout
of 0, or sread is unblocked because the thread is signaled using
some other mechanism.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>