prov/verbs: Accept FI_REMOTE_COMPLETE in fi_tx_attr::op_flags
This flag is required by the ft_finalize() function in fabtests.

FI_REMOTE_COMPLETE is essentially the behavior specified by the
InfiniBand specification for reliable connections (so the user gets this
behavior whether he asks for it or not).  To complicate things, the
iWARP DDP specification does *not* specify this behavior, so we must
reject the operation flag if it is supplied in fi_getinfo() for an iWARP
RNIC.  Note that we don't do any validity checking if this flag is
supplied in a later call to fi_sendmsg(), fi_writemsg(), etc. or when
creating an endpoint; the user gets what he deserves if doesn't check if
the operation flag is supported before using it.

If the user requests FI_REMOTE_COMPLETE, this also implies FI_COMPLETION
as an operation flag even if the user didn't specifically request it.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>