Move closing of EP into free_ep_res
Relocate where an EP is closed to ensure that it is released
before attempting to close resources that it may be bound to.
This fixes a crash seen in the sockets provider where an AV
is closed while still having an EP attached.  (The sockets
provider should have detected this using reference counting,
but it's still bad form for the app to do this.)

Signed-off-by: Sean Hefty <sean.hefty@intel.com>