fabric: Remove FI_REMOTE_SIGNAL
This can be added back after the 1.0 release.  But there is no push to include
this feature as part of the API, and there are no test apps available to verify
that it is operational.  Remove this as part of scope reduction effort in order
to hit the 1.0 release target date.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>