fabric: Introduce FI_INJECT_COMPLETE completion option
By default, operations complete when the source buffer can be reused.
Define a completion flag that indicates this.  Because this corresponds
with the fi_inject call and FI_INJECT flag, name the flag
FI_INJECT_COMPLETE.  Application can use this flag to request
optimized completions, compared to FI_TRANSMIT_COMPLETE.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>