fabric: Replace CQ write functionality with wake-up
The primary purpose of adding CQ write functionality was to provide
a way to signal a thread blocked in sread().  However, providing write
functionality ends up impacting the performance of fast path read
operations.  Rather than supporting a general write operation, replace
it with a simpler signal call that can be used to wake-up a thread
blocked in sread.

This avoids having each app implement something similar to a socketpair
that is used mainly to wake-up a thread blocked in sread, or having the
app sleep in sread for a fixed amount of time in a loop.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>