fabric: Remove duplicate ordering attributes from fi_ep_attr
The comp_order and msg_order attributes are duplicated between ep_attr,
tx_attr, and rx_attr.  Keep only the tx_attr and rx_attr definitions,
which are more precise, and eliminate the duplicate ep_attr fields.

As a result of this change, the definition of the ordering fields are
clarified as to how they apply to transmits versus receives.