prov/verbs: Fix bug where op_flags incorrectly set
Commit a819b7ec08c3c589cecb4849f3883dd0f4b4f062 introduced the
tx_op_flags variable to support FI_INJECT and set the
verbs_tx_attr.op_flags field to contain all supported tx_op_flags.
This was only to check whether or not the set of operation flags
requested in hints was supported; the flags would get overwritten with
the subset requested in hints.

However, when creating the info structure for a passive endpoint (with
src_addr but not dest_addr specified in hints), we exit out of
fi_ibv_fill_info_attr early, and never overwrite the op_flags value, so
it contains all supported attributes whether the user requested them or
not!

This commit fixes this issue by setting the verbs_tx_attr.op_flags to 0
and using the VERBS_TX_OP_FLAGS macro directly when checking the value
in hints, which has the added benefit of being more intuitive.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>