prov/psm: fix a hanging in the RMA write long protocol
The RMA write long protocol sends an AM request followed by a
tagged message. Somehow the tagged send would hang if the number
of arguments passed to the AM request is greater than 5 and the
destination endpoint is on the same node (i.e. the message goes
through the shared memory channel). This is the case when the
write has an immediate data. The fix changes the protocol to
send the immediate data as payload instead of an argument.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>