atomic_t: attempt to catch missed initializations
When !defined(HAVE_ATOMICS), fi.h falls back to using "fastlock"s to
implement the atomicity.  If those fastlocks are implemented with
pthread spinlocks on x86_64 Linux then failing to initialize the
spinlock can lead to an infinite hang.  This already happened in the
usnic provider (see PR #870).

The debug error check in this commit is an imperfect scheme, but it
should catch the most common case where a caller callocs the storage
containing the atomic_t.

Signed-off-by: Dave Goodell <dgoodell@cisco.com>