fabric/tagged: Remove tsearch call
Tsearch is a goofy, odd-ball function that pretends to be both
a synchronous call and an asynchronous call, and just all around
makes about as much sense as wearing a scarf in the summer.

Instead, define the behavior of using FI_PEEK with a receive
operation.  Since receive is sill asynchronous, FI_PEEK will
complete immediately, whether or not any data is available.
This has the added benefit that an app _could_ receive some
portion of the data (subject to the provider's implementation).
If no buffer is available for the peek to insert data into or
the provider cannot provide the data (e.g. the data could be
buffered at the sender side), the user can still obtain the
metadata (length, tag, source, etc.) related to the message.