fabric: Allow an active endpoint to inherit properties of a passive EP
To support sockets, we need a mechanism to reserved transport and
networking addresses that can be used either on the active or passive
side of a connection.  Expand the definition of a passive endpoint
to support this concept.  A passive endpoint may be assigned an
address, which is later transferred to an active endpoint using the
fi_info handle field.  Conceptually, the passive EP is 'converted'
into an active EP.  The active EP inherits all the properties that
have been set for the passive EP.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>