fabric: Change how memory registration support is exposed
Memory registration is currently exposed using 5 MR bits, including
caps, flags, and mode bits: FI_MR_OFFSET, FI_MR_KEY, FI_DYNAMIC_MR,
FI_LOCAL_MR, and FI_PROV_MR_ATTR.  Because of the combinations, this
can result in branches in the provider in data transfer operations.
For example, the provider needs to check if a key corresponds to
a virtual address or an offset based address.  It may also need to
check if the memory is pinned or might reference a dynamic MR.

To simplify the usage for both the application and the provider,
we define a new domain attribute field.  The new field is an enum
that conveys how memory registration works for the provider at a
broad level.

All MR flags, caps, and mode bits are removed, with the exception
of the FI_LOCAL_MR mode, which is still desirable for providers
to expose.  The new enum values are:

basic - This corresponds to legacy support for current IB and iWarp
hardware.  Registrations occurs on allocated memory buffers, with the
provider returning the MR attribute values, and buffers are accessed
using their virtual address + key.

scalable - This corresponds to registration desired by HPC
applications.  Registration occurs over a range of virtual memory
addresses that do not necessarily need to be back by physical pages,
with the application selecting the MR keys, and regions are accessed
using an offset + key.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>