prov/psm: fix a hanging in fi_wait
A progress thread is created when blocking wait function is called.
The progress thread is activated by a conditional variable. Due to the
timing issue, it is possible that the conditional variable be signaled
before it is waited on. That causes hanging. The fix checks the actual
condition before waiting on the conditional variable.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>