runfabtests.sh: quote grep args for zsh
If zsh is used as the default shell and the user has `setopt
extendedglob` in their settings, this `grep` command will generate some
very different output than expected, breaking the rest of the pipeline.
The "^" is the problem in this particular instance.  Single quoting the
argument fixes the issue.

Also, kill the incorrect redirection to /dev/null.  That pipeline
shouldn't be generating stdout messages anyway, and we probably *do*
want to see stderr messages (that's how I found this issue).

Signed-off-by: Dave Goodell <dgoodell@cisco.com>