fabric: Return 0 on successful asynchronous AV insertions
The man page is confusing on what value to return when
inserting addresses asynchronously into an AV, and as a
result the providers differ on their implementation.

Simplify this by having the AV insert return 0 on success
for asynchronous operations.  This makes it consistent with
other async calls.  Update the man page and providers to
handle this.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>