prov/verbs: Modify init to avoid prints from librdmacm
If the librdmacm loads but cannot find a kernel driver, it
generates an error to stderr.  Work-around this in the OFI
verbs provider to check libibverbs first.  If no verbs
devices are found, the librdmacm is not invoked.  In this
case, fail with FI_ENODATA to allow other providers to
run.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>