prov/verbs: Search for verbs device on init
The usnic device plugs into the verbs framework, but doesn't really
support verbs.  As a result, the libibverbs and librdmacm libraries
will fail to use it properly.  In particular, the librdmacm will
detect the error and send an error message to stderr.  In order to
avoid the error message, have the verbs provider search for valid
libibverbs devices.  If any are found, initialization proceeds as
normal.  If not, the verbs initialization will return ENODATA and
abort.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>