fabric: Limit FI_ATOMIC_READ to atomic::readwrite
FI_ATOMIC_READ is defined for both fi_atomic and
fi_fetch_atomic operations.  In the first case, fi_atomic
maps to an atomic::write() call.  This causes confusion
and results in branches in the provider.  The problem is
that the parameters passed into write() are used as the
source buffer in all cases, except for ATOMIC_READ, which
needs it as a destination.  Plus it's just odd that a
write() call results in a read...

Fix this by documenting that FI_ATOMIC_READ is only a
fetch atomic operation.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>