Makefile.am: ensure usnic provider has libibverbs
As noted in #1007, if you

```
$ ./configure --enable-usnic=yes --enable-verbs=dl
```

Then the usnic provider in libfabric does not end up linked against
libibverbs.  Similarly, if you:

```
$ ./configure --enable-usnic=dl --enable-verbs=dl
```

Then the usnic provider DL does not end up linked against libibverbs.

Either way the usnic provider fails, because it calls
`ibv_register_driver()`.

This commit fixes that problem by ensuring that if the verbs provider
is to be built, if we're building either the verbs or usnic provider
as a DL, then ensure we link against libibverbs.

Fixes #1007