make dist: allow a list of files to bypass git clean checks
The environment variable LIBFABRIC_DISTSCRIPT_DIRTY_FILES can now be
used to specify a whitespace-delimited list of files that will be
allowed to be git dirty during "make dist[check]".  This allows the
making of local distributions (mainly used for internal testing) with
some dirty files, but still get the safety of the rest of the git
cleanliness check.

Concrete example: local automated regression testing sets a release
number in libfabric.spec.in.  That file is therefore ok to be dirty in
this case -- the resulting tarball/RPM is only intended for local
internal testing.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>