ud_pingpong: add a coarse-grained timeout
Add "-t X" CLI option to specify a coarse-grained timeout in units of
seconds.

If the time taken for a single receive (roughly) goes past the timeout
value, quit the test with return status of 61.  This doesn't necessarily
mean that the test failed -- it just means that it timed out.

Remember: this is a UD test, and drops can happen.  This simple
pingpong program has no retransmit functionality at all, so if there's
even a single drop, the test will hang.  Hence, add functionality to
detect when that happens, and return the 61 exit status to indicate
that the test was skipped (runfabtests.sh will interpret this as "not
run, which is effectively true).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>