sockets: improve error reports when connect fails
Improve error reporting when trying to send to a target with which
connect fails. The provider was reporting -FI_EAGAIN when the
underlying connect fails for RDM type EP.

Also, inet_ntoa() cannot be called twice in a print function. It updates
a string in static storage, and if there are two calls, the second can
overwrite the buffer - order of expression evaluation isn't defined
in the C-standard. At least one print statement was printing the wrong
source address.

Signed-off-by: Sayantan Sur <sayantan.sur@intel.com>