osx/osd: use gettimeofday() on OS X
According to a few places (including
http://stackoverflow.com/a/21352348/1120009), the use of
host_get_clock_service() on OS X is both slow and inaccurate.  It
seems that gettimeofday() is quite efficient; we should simply be
using that.  And since only the sockets provider works on OS X,
reducing the precision of the timing to microseconds (from nanoseconds
-- although it looks like host_get_clock_service() may not have
actually been returning nanosecond precision, anyway), won't matter.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>