Support custom ./configure options in rpmbuild command line.
Default to building all providers with dependencies that work for RHEL7 SLES12.

If a custom configure line is specified like:
rpmbuild --define 'configopts "--enable-psm=no"' -ba libfabric.spec

then pass it to configure and disable the BuildRequires.

If if is desired to build all providers with no dependency checks call rpmbuild with --nodeps

Signed-off-by: pmmccorm <patrick.m.mccormick@intel.com>