Export fi_var_* functions as 1.0 API.
This fixes dl-provider builds.

Also add fi_getsettings() which
will allocate and return a copy of current provider registered
environment variables with values.

The caller can free the env with fi_freesettings()

This API is not thread-safe. When/if settings can be removed
during runtime (and not just shutdown) there will have to be locking
on all this stuff.

Signed-off-by: pmmccorm <patrick.m.mccormick@intel.com>