Rename fabtest to fi_ubertest
It's confusing having a test named fabtest within the fabtests
tree.  It's not clear when someone is referring to the specific
fabtest test or all of the fabtests.  Rename fabtest to
fi_ubertest, which is a waaay cooler name, to avoid confusion.

Rename complex/fabtest.c to complex/ft_main.c.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>