Fix inconsistency between send and recv paths when using FI_MSG_PREFIX.
The "usnic" provider is the only provider currently known to support the
FI_MSG_PREFIX mode. Unfortunately, it implements the feature in a confusing and
inconsistent manner. The send/recv paths assume the given user pointer points
to different portions of the buffer (the payload and the prefix, respectively).
This commit:

- Unifies this support to consistently interpret the user pointer as pointing at
  the start of the prefix region
- Maintains binary compatibility for existing compiled apps by changing the
  value of the FI_MSG_PREFIX flag and adding a new flag
  (FI_MSG_PREFIX_DEPRECATED) that has the old flag value. Source compatibility
  is not maintained, with the intent of eventually reclaiming the old flag value
  at a later release.
- Implements support for both modes of operation in the usnic provider

An alternative fix would be to use the FI_VERSION mechanism.

At this time, the only known applications that use FI_MSG_PREFIX are the
fi_ud_pingpong test and the usnic BTL in Open MPI. Maintaining binary
compatibility for Open MPI is desirable, but source compatibility is not
necessary.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>