Fix inconsistency between send and recv paths when using FI_MSG_PREFIX.
The "usnic" provider is the only provider currently known to support the
FI_MSG_PREFIX mode. Unfortunately, it implements the feature in a confusing and
inconsistent manner. The send/recv paths assume the given user pointer points
to different portions of the buffer (the payload and the prefix, respectively).
This commit:

- Unifies this support to consistently interpret the user pointer as pointing at
  the start of the prefix region
- Maintains binary compatibility for existing applications by changing the
  API version to (1, 1).
- Source compatibility is maintained provided that applications are reporting
  use of API version (1, 0).
- Adds code the usnic provider to correctly handle both (1,0) and (1,1)
  behavior.

At this time, the only known applications that use FI_MSG_PREFIX are the
fi_ud_pingpong test and the usnic BTL in Open MPI.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>