prov/all: Report version that is actually supported
The providers, even if built-in, should only report the
version of the interfaces that they are prepared to support.
They are currently in sync with the latest version of the
framework, but this may not always be the case.

The concern here is that a vendor may extract their provider
code from the source tree and ship it separately from libfabric.
If they do so, their code must reflect the version that they
are coded to handle, rather than picking up the version of
the installed library.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>