prov/sockets: Check for disconnected sockets
If a socket is disconnected by the remote peer, select() will
return POLLIN, indicating that reading from the socket will not
block.  A recv() call against the socket will return 0 in this
case.

The progress code currently ignores disconnect, and as a result
will busy spin checking for receive data on the socket.  This
can result in significant delays during shutdown, since the
progress thread never stops executing, holding locks for an
extended period of time.

Add a disconnect state to the sock_conn structure.  When a
socket has been disconnected, mark it as such and skip all
future checks trying to read from it.  This will allow the
progress thread to wait on only those sockets that are still
active.

Signed-off-by: Jithin Jose <jithin.jose@intel.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>