usnic: Update usdf_dgram_prefix_send.
The previous implementation of usdf_dgram_prefix_send expected a length
parameter equal to the length of the payload rather than the full send
buffer. This commit changes the behavior of usdf_dgram_prefix_send to
expect a length parameter that includes the prefix size in the length
and updates the headers appropriately.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>