man: remove prefix 8-byte alignment requirement
This places an unnecessary burden on providers when implementing the RX
path for FI_MSG_PREFIX mode.  For example, the usnic provider uses a
42-byte UDP/IP/Ethernet header for its EP_DGRAM packets, but 42 is not a
multiple of 8 (or 4).  Requiring rounding this prefix size up to 48 (or
even 44) would work, but either wastefully transmits extra bytes or
requires additional lookups and multilevel pointer dereferences on RX to
determine whether the received completion arrived on an RQ that is in
FI_MSG_PREFIX mode.

Presumably the requirement was put in place to make it easier for
applications to align their payload buffers to 8-byte boundaries and
accelerate app-level memory copies?

Signed-off-by: Dave Goodell <dgoodell@cisco.com>