prov/sockets: Do not ignore dest_addr in hints
Commit cad9e622b3b67d05031e8004e673161685449db6 introduced a regression
when the user ran the following sequence of code:

> hints->dest_addr = address;
> hints->dest_addrlen = address_len;
> ret = fi_getinfo(EXS_FI_VERSION, NULL, NULL, 0, &hints, all_info);

The dest_addr supplied by the user is ignored and replaced with
the address corresponding to the hostname for both source and
destination address.  In my cluster setup, the RDMA interfaces use a
different naming scheme than the Ethernet management interfaces, so this
causes the sockets provider to use the wrong interface for establishing
connections.

This commit fixes the issue by not implicitly using the hostname if the
user in fact specified a destination address in hints.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>