prov/verbs: Set fabric_attr->prov_name in connreq info structure
Commit 69776783ff03 ("prov/verbs: Improve getinfo attributes check")
introduced a regression which unset the prov_name field from the
info->fabric_attr structure for incoming connection requests.

Although the libfabric core automatically sets fabric_attr->prov_name on
fi_getinfo(), it cannot do this for the info objects created for an
incoming connection request.  Thus, the verbs provider needs to do this
manually.

This issue was uncovered because UNH EXS makes no assumption about what
the fabric object will be when creating a domain, and thus creates a new
fabric handle if one is not supplied in fabric_attr->fabric.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>