usnic: Make send completion lengths reflect sizes given to library.
The send routines should generate completions with a length field that
matches the size given as input. This commit fixes the reported sizes to
comply with the API behavior by setting wp_len to the size of the
payload in all send cases and adjusting for send completions upon
polling a completion queue.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>