prov/verbs: Implement fi_setname and pep migration to active ep.
fi_endpoint will now use passive EP's rdma_cm_id if its fid is
passed in info->handle.

fi_setname can now be ran on a passive EP to bind it to a specific address.
If the passive EP has already been bound the a rdma_cm_id, the id will be
destroyed and a new one will be created and the address passed to fi_setname
will be bound to the new id.

Similarly, fi_setname for an active endpoint will destroy the previous
rdma_cm_id and create a new one.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>