complex/fi_ubertest: Add support for config files.
This patch adds support for complex/fi_ubertest to read test configuration
from a file. The configuration should be in JSON format. An example
config file is also added.

"jsmn", a 3rd party JSON parser that is freely available under MIT license
is used. The "jsmn" parser reads a JSON formatted file and returns the JSON
tokens which is then organized into an array of ft_set structures.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>