simple/msg_sockets: Remove infiniband/ib.h include
Since fabtests needs to work on an IB-less system, and AF_IB is not yet
supported by the verbs provider, just remove the AF_IB support in the
test for now.  It can be added back in a more robust way once the verbs
provider gains support for AF_IB addresses.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>