spec: Remove BuildRequires dependencies from spec file
It is extremely unlikely that many users will have a
system with all devices supported by libfabric present.
Avoid build requirements in the spec file.  Instead allow the
configure script to automatically detect what can be built.
This fixes a build error when integrating libfabric with the
OFED build environment.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>