cron-make-nightly-tarball: Install into temporary directory
We need to install libfabric somewhere so that fabtests can build
against it.  We choose to do it into a temporary directory since there
is no real reason for it to persist on flatbed, and it ensures we get a
clean install directory with the latest libfabric every time the script
runs, which is important when the libfabric API is extended and fabtests
immediately requires the new libfabric.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>