cron-make-nightly-tarball: Also make/submit fabtests tarballs
We currently are not running fabtests through Coverity, and there is
really no good reason for this.  This commit updates the nightly tarball
script to submit fabtests to Coverity Scan nightly.  As a side effect,
we now also generate nightly tarballs for fabtests which are accessible
from the same downloads directory as the libfabric tarballs.

Suggested-by: Ben Turrubiates <bturrubi@cisco.com>
Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>