Correct error strings on EAI_SYSTEM
The error return EAI_SYSTEM for getaddrinfo() and getnameinfo()
indicates that the user should check errno for the cause of the error.
This test previously did the wrong thing everywhere that EAI_SYSTEM could
be returned, and this commit fixes the code to print the correct error
string using strerror(errno).

This commit fixes Coverity CID 131051.

Signed-off-by: Patrick MacArthur <pmacarth@iol.unh.edu>